Product and Blog Paths
Almost done with the productController file. Need to tweak the code to accept Images for the "imagePath", then it will basically be done in the data-pipeline aspect.

Blog path is essentially done, minus the addition of two new paths that are important to the new Comment Schema and their relation to the blogs.